+++ client/next-app/src/app/blog/page.js
@@ -1,11 +1,11 @@
-import Image from 'next/image'
-import Link from 'next/link'
-import NavBar from '../components/ui/NavBar'
+import Image from "next/image";
+import Link from "next/link";
+import { NavBar } from "@/components/NavBar";
 
 export default function Blog() {
   return (
-    <main>
+    <div>
       <NavBar></NavBar>
-    </main>
-  )
+    </div>
+  );
 }
